- Store price as double instead of integer - Start inplementation of price calculation: Sum all prices within a group, in case the price is larger than 0 and the same for the quantity, and then sum up all group prices. - Display total price in the info panel

+++ src/indico/checkin/core/api/IndicoFullRegistrantParser.java
@@ -114,7 +114,7 @@ public class IndicoFullRegistrantParser {
 		if(!item.get("id").toString().isEmpty())
 			field.setId(Long.parseLong(item.get("id").toString()));
 		if(!item.get("price").toString().isEmpty())
-			field.setPrice(Integer.parseInt(item.get("price").toString()));
+			field.setPrice(Double.parseDouble(item.get("price").toString()));
 		if(!item.get("quantity").toString().isEmpty())
 			field.setQuantity(Integer.parseInt(item.get("quantity").toString()));
 		field.setCurrency((String)item.get("currency"));
